move method into test case
