+++ src/main/webapp/app/shared/util/utils.ts
@@ -4,26 +4,6 @@ import { Result } from 'app/entities/result.model';
 import { Course } from 'app/entities/course.model';
 import { Exercise } from 'app/entities/exercise.model';
 
-// Cartesian product helper function
-const cartesianConcatHelper = (a: any[], b: any[]): any[][] => ([] as any[][]).concat(...a.map((a2) => b.map((b2) => ([] as any[]).concat(a2, b2))));
-
-/**
- * Returns the cartesian product for all arrays provided to the function.
- * Type of the arrays does not matter, it will just return the combinations without any type information.
- * Implementation taken from here: https://gist.github.com/ssippe/1f92625532eef28be6974f898efb23ef.
- * @param a an array
- * @param b another array
- * @param c rest of arrays
- */
-export const cartesianProduct = (a: any[], b: any[], ...c: any[][]): any[] => {
-    if (!b || b.length === 0) {
-        return a;
-    }
-    const [b2, ...c2] = c;
-    const fab = cartesianConcatHelper(a, b);
-    return cartesianProduct(fab, b2, ...c2);
-};
-
 export function cleanString(str?: string): string {
     if (!str) {
         return '';
@@ -56,7 +36,7 @@ export const stringifyCircular = (val: any): string => {
  * @param val Object to stringify
  * @param ignoredFields Fields to omit from object before converting to string
  */
-export const stringifyIgnoringFields = (val: any, ...ignoredFields: string[]): string => {
+export const stringifyIgnoringFields = <T extends object>(val: T, ...ignoredFields: string[]): string => {
     return JSON.stringify(omit(val, ignoredFields));
 };
